SSA-895 To implement Tempo settings screen SSA-897 Time picker dialog.

+++ MiniumApps/launcher3/src/main/java/co/siempo/phone/service/StatusBarService.java
@@ -8,7 +8,6 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.SharedPreferences;
-import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageManager;
 import android.database.ContentObserver;
 import android.hardware.Camera;
@@ -16,6 +15,7 @@ import android.hardware.camera2.CameraAccessException;
 import android.hardware.camera2.CameraManager;
 import android.net.Uri;
 import android.os.Build;
+import android.os.CountDownTimer;
 import android.os.Handler;
 import android.os.IBinder;
 import android.provider.ContactsContract;
@@ -24,17 +24,23 @@ import android.support.v4.content.ContextCompat;
 import android.text.TextUtils;
 import android.util.Log;
 
-import java.util.List;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Calendar;
+import java.util.Timer;
+import java.util.TimerTask;
 
-import co.siempo.phone.db.DBClient;
 import co.siempo.phone.R;
+import co.siempo.phone.db.DBClient;
 import co.siempo.phone.event.TorchOnOff;
 import co.siempo.phone.helper.FirebaseHelper;
+import co.siempo.phone.util.PackageUtil;
 import de.greenrobot.event.EventBus;
 import de.greenrobot.event.Subscribe;
 import minium.co.core.app.CoreApplication;
 import minium.co.core.event.AppInstalledEvent;
 import minium.co.core.event.FirebaseEvent;
+import minium.co.core.log.Tracer;
 
 import static co.siempo.phone.SiempoNotificationBar.NotificationUtils.ANDROID_CHANNEL_ID;
 
@@ -44,27 +50,41 @@ import static co.siempo.phone.SiempoNotificationBar.NotificationUtils.ANDROID_CH
 
 public class StatusBarService extends Service {
 
+    public static boolean isFlashOn = false;
+    SharedPreferences sharedPreferences;
+    Timer timer;
+    MyTimerTask timerTask;
+    CountDownTimer countDownTimer;
+    Context context;
+    ArrayList<Integer> everyHourList = new ArrayList<>();
+    ArrayList<Integer> everyTwoHourList = new ArrayList<>();
+    ArrayList<Integer> everyFourHoursList = new ArrayList<>();
     private CameraManager cameraManager;
     private String mCameraId;
     @SuppressWarnings("deprecation")
     private Camera camera;
     @SuppressWarnings("deprecation")
     private Camera.Parameters parameters;
-    public static boolean isFlashOn = false;
-
-    SharedPreferences sharedPreferences;
     private MyObserver myObserver;
     private AppInstallUninstall appInstallUninstall;
 
+    public StatusBarService() {
+    }
 
     @Override
     public void onCreate() {
         super.onCreate();
+        context = this;
         sharedPreferences = getSharedPreferences("DroidPrefs", 0);
         registerObserverForContact();
         registerObserverForAppInstallUninstall();
         EventBus.getDefault().register(this);
-
+        everyHourList.addAll(Arrays.asList(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24));
+        everyTwoHourList.addAll(Arrays.asList(1, 3, 5, 7, 9, 11, 13, 15, 17, 19, 21, 23));
+        everyFourHoursList.addAll(Arrays.asList(1, 4, 8, 12, 16, 20, 24));
+        timerTask = new MyTimerTask();
+        timer = new Timer();
+        timer.schedule(timerTask, 0, 60000);
     }
 
     @Override
@@ -96,7 +116,6 @@ public class StatusBarService extends Service {
     }
 
 
-
     /**
      * Observer for when new contact adding or updating any exiting contact.
      */
@@ -116,10 +135,6 @@ public class StatusBarService extends Service {
         }
     }
 
-
-    public StatusBarService() {
-    }
-
     @Subscribe
     public void tourchOnOff(TorchOnOff torchOnOFF) {
         if (torchOnOFF.isRunning()) {
@@ -133,7 +148,6 @@ public class StatusBarService extends Service {
     public void firebaseEvent(FirebaseEvent firebaseEvent) {
         FirebaseHelper.getIntance().logScreenUsageTime(firebaseEvent.getScreenName(), firebaseEvent.getStrStartTime());
     }
-
 
     @Override
     public IBinder onBind(Intent intent) {
@@ -211,6 +225,63 @@ public class StatusBarService extends Service {
         isFlashOn = false;
     }
 
+    @Override
+    public void onDestroy() {
+        EventBus.getDefault().unregister(this);
+        if (myObserver != null)
+            getContentResolver().unregisterContentObserver(myObserver);
+        if (appInstallUninstall != null)
+            unregisterReceiver(appInstallUninstall);
+        super.onDestroy();
+    }
+
+    class MyTimerTask extends TimerTask {
+        public void run() {
+            if (PackageUtil.isSiempoLauncher(context)) {
+                Calendar calendar = Calendar.getInstance();
+                int systemHours = calendar.get(Calendar.HOUR_OF_DAY);
+                int systemMinutes = calendar.get(Calendar.MINUTE);
+                if ((sharedPreferences.getInt("tempoType", 0) == 1)) {
+                    int batchTime = sharedPreferences.getInt("batchTime", 15);
+                    if (batchTime == 15) {
+                        if (systemMinutes == 0 || systemMinutes == 15 || systemMinutes == 30 || systemMinutes == 45) {
+                            Tracer.d("Batch::" + "15 minute interval");
+                        }
+                    } else if (batchTime == 30) {
+                        if (systemMinutes == 0 || systemMinutes == 30) {
+                            Tracer.d("Batch::" + "15 minute interval");
+                        }
+                    } else if (batchTime == 1) {
+                        if (everyHourList.contains(systemHours) && systemMinutes == 0) {
+                            Tracer.d("Batch::" + "Every Hour interval");
+                        }
+                    } else if (batchTime == 2) {
+                        if (everyTwoHourList.contains(systemHours) && systemMinutes == 0) {
+                            Tracer.d("Batch::" + "Every 2 Hour interval");
+                        }
+                    } else if (batchTime == 4) {
+                        if (everyFourHoursList.contains(systemHours) && systemMinutes == 0) {
+                            Tracer.d("Batch::" + "Every 4 Hour interval");
+                        }
+                    }
+
+                } else if ((sharedPreferences.getInt("tempoType", 0) == 2)) {
+                    String strTimeData = sharedPreferences.getString("onlyAt", "");
+                    if (!strTimeData.equalsIgnoreCase("")) {
+                        String strTime[] = strTimeData.split(",");
+                        for (String str : strTime) {
+                            int hours = Integer.parseInt(str.split(":")[0]);
+                            int minutes = Integer.parseInt(str.split(":")[1]);
+                            if (hours == systemHours && minutes == systemMinutes) {
+                                Tracer.d("Only at::" + str);
+                            }
+                        }
+                    }
+                }
+            }
+
+        }
+    }
 
     private class MyObserver extends ContentObserver {
         MyObserver(Handler handler) {
@@ -241,17 +312,20 @@ public class StatusBarService extends Service {
                 if (intent != null && intent.getAction() != null) {
                     if (intent.getAction().equals(Intent.ACTION_PACKAGE_ADDED)) {
                         String installPackageName;
-                        installPackageName = intent.getData().getEncodedSchemeSpecificPart();
-                        Log.d("Testing with device.", "Added" + installPackageName);
+                        if (intent.getData().getEncodedSchemeSpecificPart() != null) {
+                            installPackageName = intent.getData().getEncodedSchemeSpecificPart();
+                            Log.d("Testing with device.", "Added" + installPackageName);
+                        }
 
                     } else if (intent.getAction().equals(Intent.ACTION_PACKAGE_REMOVED)) {
                         String uninstallPackageName;
-                        uninstallPackageName = intent.getData().getSchemeSpecificPart();
-                        Log.d("Testing with device.", "Removed" + uninstallPackageName);
-                        if (!TextUtils.isEmpty(uninstallPackageName)) {
-                            new DBClient().deleteMsgByPackageName(uninstallPackageName);
+                        if (intent.getData().getEncodedSchemeSpecificPart() != null) {
+                            uninstallPackageName = intent.getData().getSchemeSpecificPart();
+                            Log.d("Testing with device.", "Removed" + uninstallPackageName);
+                            if (!TextUtils.isEmpty(uninstallPackageName)) {
+                                new DBClient().deleteMsgByPackageName(uninstallPackageName);
+                            }
                         }
-
                     }
                     sharedPreferences.edit().putBoolean("isAppUpdated", true).apply();
                     EventBus.getDefault().post(new AppInstalledEvent(true));
@@ -262,17 +336,5 @@ public class StatusBarService extends Service {
             }
 
         }
-    }
-
-
-
-    @Override
-    public void onDestroy() {
-        EventBus.getDefault().unregister(this);
-        if (myObserver != null)
-            getContentResolver().unregisterContentObserver(myObserver);
-        if (appInstallUninstall != null)
-            unregisterReceiver(appInstallUninstall);
-        super.onDestroy();
     }
 }
